+++ packages/pwa-kit-dev/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## v3.16.0-dev (Dec 17, 2025)
-- Update jest, archiver and remove rimraf [#3663](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3663)
+- Update jest, archiver and remove rimraf. Add `customExportConditions` to Jest config for jsdom compatibility with ESM packages [#3663](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3663)
 - Add Node 24 support, remove legacy `url` module import. Drop Node 16 support [#3652](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3652)
 
 ## v3.15.0 (Dec 17, 2025)